+++ jest.config.js
@@ -4,7 +4,7 @@ module.exports = {
   modulePathIgnorePatterns: ['dist'],
   testMatch: ['<rootDir>/packages/**/*.spec.[jt]s?(x)'],
   moduleNameMapper: {
-    '^@web3-wallet/(.*)$': [
+    '^@react-web3-wallet/(.*)$': [
       '<rootDir>/packages/$1/src',
       '<rootDir>/connectors/$1/src',
     ],