+++ packages/react/src/hooks/balance.ts
@@ -1,11 +1,10 @@
-import type { BigNumber } from '@ethersproject/bignumber';
 import { useEffect, useState } from 'react';
 
 import type { Wallet } from '../types';
 
 export type BalanceHooks = {
-  useBalances: () => BigNumber[];
-  useBalance: () => BigNumber | undefined;
+  useBalances: () => bigint[];
+  useBalance: () => bigint | undefined;
 };
 
 export const createBalanceHooks = (wallet: Wallet): BalanceHooks => {
@@ -14,7 +13,7 @@ export const createBalanceHooks = (wallet: Wallet): BalanceHooks => {
   const useBalances: BalanceHooks['useBalances'] = () => {
     const provider = useProvider();
     const accounts = useAccounts();
-    const [balances, setBalances] = useState<BigNumber[]>([]);
+    const [balances, setBalances] = useState<bigint[]>([]);
 
     useEffect(() => {
       let canceled = false;